[trie] Refactor trie ops related code
