+++ core/store/src/trie/mem/updating.rs
@@ -1,211 +1,71 @@
-//! Structures and logic for updating in-memory trie.
-//!
-//! DISCLAIMER: This is in process of rewriting to generic structures.
-//! See #12324.
-//! For now we keep the old types together with new ones to change logic
-//! incrementally (for example, `GenericNodeOrIndex` and `OldOrUpdatedNodeId`).
-//! New methods will be prefixed with `generic_` to distinguish them from the
-//! old ones. When the old methods are removed, the prefix will be dropped.
+use std::collections::{BTreeMap, HashMap};
+use std::sync::Arc;
+
+use near_primitives::errors::StorageError;
+use near_primitives::hash::{hash, CryptoHash};
+use near_primitives::state::FlatStateValue;
+
+use crate::trie::mem::node::MemTrieNodeId;
+use crate::trie::ops::insert_delete::GenericTrieUpdateInsertDelete;
+use crate::trie::ops::interface::{
+    GenericNodeOrIndex, GenericTrieUpdate, GenericTrieValue, GenericUpdatedNodeId,
+    GenericUpdatedTrieNode, GenericUpdatedTrieNodeWithSize,
+};
+use crate::trie::{Children, MemTrieChanges, TrieRefcountDeltaMap, TRIE_COSTS};
+use crate::{RawTrieNode, RawTrieNodeWithSize, TrieChanges};
 
 use super::arena::{ArenaMemory, ArenaMut};
 use super::flexible_data::children::ChildrenView;
 use super::metrics::MEM_TRIE_NUM_NODES_CREATED_FROM_UPDATES;
-use super::node::{InputMemTrieNode, MemTrieNodeId, MemTrieNodeView};
-use crate::trie::insert_delete::NodesStorage;
-use crate::trie::{
-    Children, MemTrieChanges, StorageValueHandle, TrieRefcountDeltaMap, ValueHandle, TRIE_COSTS,
-};
-use crate::{NibbleSlice, RawTrieNode, RawTrieNodeWithSize, TrieChanges};
-use near_primitives::errors::StorageError;
-use near_primitives::hash::{hash, CryptoHash};
-use near_primitives::state::{FlatStateValue, GenericTrieValue, ValueRef};
-use std::collections::{BTreeMap, HashMap};
-use std::sync::Arc;
-
-/// For updated nodes, the ID is simply the index into the array of updated nodes we keep.
-pub type GenericUpdatedNodeId = usize;
+use super::node::{InputMemTrieNode, MemTrieNodeView};
 
 pub type UpdatedMemTrieNodeId = usize;
 
-/// An old node means a node in the current in-memory trie. An updated node means a
-/// node we're going to store in the in-memory trie but have not constructed there yet.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum GenericNodeOrIndex<GenericTrieNodePtr> {
-    Old(GenericTrieNodePtr),
-    Updated(GenericUpdatedNodeId),
-}
-
 pub type OldOrUpdatedNodeId = GenericNodeOrIndex<MemTrieNodeId>;
-
-/// Trait for trie values to get their length.
-pub trait HasValueLength {
-    fn len(&self) -> u64;
-}
-
-impl HasValueLength for FlatStateValue {
-    fn len(&self) -> u64 {
-        self.value_len() as u64
-    }
-}
-
-impl HasValueLength for ValueHandle {
-    fn len(&self) -> u64 {
-        match self {
-            ValueHandle::HashAndSize(value) => value.length as u64,
-            ValueHandle::InMemory(value) => value.1 as u64,
-        }
-    }
-}
-
-/// An updated node - a node that will eventually become a trie node.
-/// It references children that are either old or updated nodes.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum GenericUpdatedTrieNode<GenericTrieNodePtr, GenericValueHandle> {
-    /// Used for either an empty root node (indicating an empty trie), or as a temporary
-    /// node to ease implementation.
-    Empty,
-    Leaf {
-        extension: Box<[u8]>,
-        value: GenericValueHandle,
-    },
-    Extension {
-        extension: Box<[u8]>,
-        child: GenericNodeOrIndex<GenericTrieNodePtr>,
-    },
-    /// Corresponds to either a Branch or BranchWithValue node.
-    Branch {
-        children: Box<[Option<GenericNodeOrIndex<GenericTrieNodePtr>>; 16]>,
-        value: Option<GenericValueHandle>,
-    },
-}
 
 pub type UpdatedMemTrieNode = GenericUpdatedTrieNode<MemTrieNodeId, FlatStateValue>;
 
-/// An updated node with its memory usage.
-/// Needed to recompute subtree function (memory usage) on the fly.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct GenericUpdatedTrieNodeWithSize<GenericTrieNodePtr, GenericValueHandle> {
-    pub node: GenericUpdatedTrieNode<GenericTrieNodePtr, GenericValueHandle>,
-    pub memory_usage: u64,
-}
-
-impl<N, V> GenericUpdatedTrieNodeWithSize<N, V> {
-    pub fn empty() -> Self {
-        Self { node: GenericUpdatedTrieNode::Empty, memory_usage: 0 }
-    }
-}
-
-impl<GenericTrieNodePtr, FlatStateValue: HasValueLength>
-    GenericUpdatedTrieNode<GenericTrieNodePtr, FlatStateValue>
-{
-    fn memory_usage_value(value_length: u64) -> u64 {
-        value_length * TRIE_COSTS.byte_of_value + TRIE_COSTS.node_cost
-    }
-
-    /// Returns the memory usage of the **single** node, in Near's trie cost
-    /// terms, not in terms of the physical memory usage.
-    pub fn memory_usage_direct(&self) -> u64 {
-        match self {
-            Self::Empty => {
-                // DEVNOTE: empty nodes don't exist in storage.
-                // In the in-memory implementation Some(TrieNode::Empty) and None are interchangeable as
-                // children of branch nodes which means cost has to be 0
-                0
-            }
-            Self::Leaf { extension, value } => {
-                TRIE_COSTS.node_cost
-                    + (extension.len() as u64) * TRIE_COSTS.byte_of_key
-                    + Self::memory_usage_value(value.len())
-            }
-            Self::Branch { value, .. } => {
-                TRIE_COSTS.node_cost
-                    + value.as_ref().map_or(0, |value| Self::memory_usage_value(value.len()))
-            }
-            Self::Extension { extension, .. } => {
-                TRIE_COSTS.node_cost + (extension.len() as u64) * TRIE_COSTS.byte_of_key
-            }
-        }
-    }
-}
-
 pub type UpdatedMemTrieNodeWithSize = GenericUpdatedTrieNodeWithSize<MemTrieNodeId, FlatStateValue>;
 
-/// Trait for trie updates to handle updated nodes.
-///
-/// So far, this is used to handle key-value insertions, deletions and range
-/// retain operation. To be performant, such logic requires keeping track of
-/// intermediate updated nodes, together with subtree function (memory usage).
-///
-/// `GenericTrieUpdate` abstracts the storage of updated nodes for the original
-/// node type `GenericTrieNodePtr`.
-///
-/// In this storage, nodes are indexed by `GenericUpdatedNodeId`.
-/// Node is stored as `GenericUpdatedTrieNodeWithSize`, which stores children
-/// as `GenericNodeOrIndex`. Each child may be either an old node or an updated
-/// node.
-///
-/// The flow of interaction with this storage is:
-/// - In the beginning, call `ensure_updated` for the
-/// `GenericNodeOrIndex::Old(root_node)` which returns `GenericUpdatedNodeId`,
-/// it should be zero.
-/// - For every update (single insert, single delete, recursive range
-/// operation...), call corresponding method with `GenericUpdatedNodeId` for
-/// the root.
-/// - Then, we hold the invariant that on every descent we have
-/// `GenericUpdatedNodeId`.
-/// - So, first, we call `take_node` to get `GenericUpdatedTrieNodeWithSize`
-/// back;
-/// - We possibly descend into its children and modify the node;
-/// - Then, we call `place_node` to put the node back and return to the
-/// node parent.
-/// - Finally, we end up with storage of new nodes, which are used to produce
-/// new state root. The exact logic depends on trait implementation.
-///
-/// TODO(#12361): instead of `GenericValueHandle`, consider always using
-/// `FlatStateValue`.
-///
-/// Note that it has nothing to do with `TrieUpdate` used for runtime to store
-/// temporary state changes (TODO(#12361) - consider renaming it).
-pub(crate) trait GenericTrieUpdate<'a, GenericTrieNodePtr, GenericValueHandle> {
-    /// If the ID was old, converts underlying node to an updated one.
-    fn ensure_updated(
-        &mut self,
-        node: GenericNodeOrIndex<GenericTrieNodePtr>,
-    ) -> Result<GenericUpdatedNodeId, StorageError>;
+impl UpdatedMemTrieNodeWithSize {
+    /// Converts an existing in-memory trie node into an updated one that is
+    /// equivalent.
+    pub fn from_existing_node_view<'a, M: ArenaMemory>(view: MemTrieNodeView<'a, M>) -> Self {
+        let memory_usage = view.memory_usage();
+        let node = match view {
+            MemTrieNodeView::Leaf { extension, value } => UpdatedMemTrieNode::Leaf {
+                extension: extension.to_vec().into_boxed_slice(),
+                value: value.to_flat_value(),
+            },
+            MemTrieNodeView::Branch { children, .. } => UpdatedMemTrieNode::Branch {
+                children: Box::new(Self::convert_children_to_updated(children)),
+                value: None,
+            },
+            MemTrieNodeView::BranchWithValue { children, value, .. } => {
+                UpdatedMemTrieNode::Branch {
+                    children: Box::new(Self::convert_children_to_updated(children)),
+                    value: Some(value.to_flat_value()),
+                }
+            }
+            MemTrieNodeView::Extension { extension, child, .. } => UpdatedMemTrieNode::Extension {
+                extension: extension.to_vec().into_boxed_slice(),
+                child: OldOrUpdatedNodeId::Old(child.id()),
+            },
+        };
+        Self { node, memory_usage }
+    }
 
-    /// Takes a node from the set of updated nodes, setting it to None.
-    /// It is expected that place_node is then called to return the node to
-    /// the same slot.
-    fn take_node(
-        &mut self,
-        node_id: GenericUpdatedNodeId,
-    ) -> GenericUpdatedTrieNodeWithSize<GenericTrieNodePtr, GenericValueHandle>;
-
-    /// Puts a node to the set of updated nodes at specific index.
-    /// Needed when reference to node from parent needs to be preserved.
-    fn place_node_at(
-        &mut self,
-        node_id: GenericUpdatedNodeId,
-        node: GenericUpdatedTrieNodeWithSize<GenericTrieNodePtr, GenericValueHandle>,
-    );
-
-    /// Puts a new node into the set of updated nodes.
-    fn place_node(
-        &mut self,
-        node: GenericUpdatedTrieNodeWithSize<GenericTrieNodePtr, GenericValueHandle>,
-    ) -> GenericUpdatedNodeId;
-
-    /// Gets a node reference in the set of updated nodes.
-    fn get_node_ref(
-        &self,
-        node_id: GenericUpdatedNodeId,
-    ) -> &GenericUpdatedTrieNodeWithSize<GenericTrieNodePtr, GenericValueHandle>;
-
-    /// Stores a state value in the trie.
-    fn store_value(&mut self, value: GenericTrieValue) -> GenericValueHandle;
-
-    /// Deletes a state value from the trie.
-    fn delete_value(&mut self, value: GenericValueHandle) -> Result<(), StorageError>;
+    fn convert_children_to_updated<'a, M: ArenaMemory>(
+        view: ChildrenView<'a, M>,
+    ) -> [Option<OldOrUpdatedNodeId>; 16] {
+        let mut children = [None; 16];
+        for i in 0..16 {
+            if let Some(child) = view.get(i) {
+                children[i] = Some(OldOrUpdatedNodeId::Old(child.id()));
+            }
+        }
+        children
+    }
 }
 
 /// Keeps values and internal nodes accessed on updating memtrie.
@@ -267,47 +127,6 @@ pub struct MemTrieUpdate<'a, M: ArenaMemory> {
     /// Tracks trie changes necessary to make on-disk updates and recorded
     /// storage.
     tracked_trie_changes: Option<TrieChangesTracker>,
-}
-
-impl UpdatedMemTrieNodeWithSize {
-    /// Converts an existing in-memory trie node into an updated one that is
-    /// equivalent.
-    pub fn from_existing_node_view<'a, M: ArenaMemory>(view: MemTrieNodeView<'a, M>) -> Self {
-        let memory_usage = view.memory_usage();
-        let node = match view {
-            MemTrieNodeView::Leaf { extension, value } => UpdatedMemTrieNode::Leaf {
-                extension: extension.to_vec().into_boxed_slice(),
-                value: value.to_flat_value(),
-            },
-            MemTrieNodeView::Branch { children, .. } => UpdatedMemTrieNode::Branch {
-                children: Box::new(Self::convert_children_to_updated(children)),
-                value: None,
-            },
-            MemTrieNodeView::BranchWithValue { children, value, .. } => {
-                UpdatedMemTrieNode::Branch {
-                    children: Box::new(Self::convert_children_to_updated(children)),
-                    value: Some(value.to_flat_value()),
-                }
-            }
-            MemTrieNodeView::Extension { extension, child, .. } => UpdatedMemTrieNode::Extension {
-                extension: extension.to_vec().into_boxed_slice(),
-                child: OldOrUpdatedNodeId::Old(child.id()),
-            },
-        };
-        Self { node, memory_usage }
-    }
-
-    fn convert_children_to_updated<'a, M: ArenaMemory>(
-        view: ChildrenView<'a, M>,
-    ) -> [Option<OldOrUpdatedNodeId>; 16] {
-        let mut children = [None; 16];
-        for i in 0..16 {
-            if let Some(child) = view.get(i) {
-                children[i] = Some(OldOrUpdatedNodeId::Old(child.id()));
-            }
-        }
-        children
-    }
 }
 
 impl<'a, M: ArenaMemory> GenericTrieUpdate<'a, MemTrieNodeId, FlatStateValue>
@@ -376,115 +195,6 @@ impl<'a, M: ArenaMemory> GenericTrieUpdate<'a, MemTrieNodeId, FlatStateValue>
                 .or_insert(1);
         }
 
-        Ok(())
-    }
-}
-
-pub(crate) type TrieStorageNodePtr = CryptoHash;
-
-pub(crate) type UpdatedTrieStorageNode = GenericUpdatedTrieNode<TrieStorageNodePtr, ValueHandle>;
-
-pub(crate) type UpdatedTrieStorageNodeWithSize =
-    GenericUpdatedTrieNodeWithSize<TrieStorageNodePtr, ValueHandle>;
-
-impl UpdatedTrieStorageNodeWithSize {
-    pub fn from_raw_trie_node_with_size(node: RawTrieNodeWithSize) -> Self {
-        Self {
-            node: UpdatedTrieStorageNode::from_raw_trie_node(node.node),
-            memory_usage: node.memory_usage,
-        }
-    }
-}
-
-impl UpdatedTrieStorageNode {
-    fn new_branch(children: Children, value: Option<ValueRef>) -> Self {
-        let children =
-            Box::new(children.0.map(|child| child.map(|id| GenericNodeOrIndex::Old(id))));
-        let value = value.map(ValueHandle::HashAndSize);
-        Self::Branch { children, value }
-    }
-
-    /// Conversion from the node just read from trie storage.
-    pub fn from_raw_trie_node(node: RawTrieNode) -> Self {
-        match node {
-            RawTrieNode::Leaf(extension, value) => Self::Leaf {
-                extension: extension.to_vec().into_boxed_slice(),
-                value: ValueHandle::HashAndSize(value),
-            },
-            RawTrieNode::BranchNoValue(children) => Self::new_branch(children, None),
-            RawTrieNode::BranchWithValue(value, children) => {
-                Self::new_branch(children, Some(value))
-            }
-            RawTrieNode::Extension(extension, child) => Self::Extension {
-                extension: extension.to_vec().into_boxed_slice(),
-                child: GenericNodeOrIndex::Old(child),
-            },
-        }
-    }
-}
-
-const INVALID_STORAGE_HANDLE: &str = "invalid storage handle";
-
-impl<'a> GenericTrieUpdate<'a, TrieStorageNodePtr, ValueHandle> for NodesStorage<'a> {
-    fn ensure_updated(
-        &mut self,
-        node: GenericNodeOrIndex<TrieStorageNodePtr>,
-    ) -> Result<GenericUpdatedNodeId, StorageError> {
-        match node {
-            GenericNodeOrIndex::Old(node_hash) => {
-                self.trie.move_node_to_mutable(self, &node_hash).map(|handle| handle.0)
-            }
-            GenericNodeOrIndex::Updated(node_id) => Ok(node_id),
-        }
-    }
-
-    fn take_node(&mut self, index: GenericUpdatedNodeId) -> UpdatedTrieStorageNodeWithSize {
-        self.nodes
-            .get_mut(index)
-            .expect(INVALID_STORAGE_HANDLE)
-            .take()
-            .expect(INVALID_STORAGE_HANDLE)
-    }
-
-    fn place_node_at(&mut self, index: GenericUpdatedNodeId, node: UpdatedTrieStorageNodeWithSize) {
-        debug_assert!(self.nodes.get(index).expect(INVALID_STORAGE_HANDLE).is_none());
-        self.nodes[index] = Some(node);
-    }
-
-    fn place_node(&mut self, node: UpdatedTrieStorageNodeWithSize) -> GenericUpdatedNodeId {
-        let index = self.nodes.len();
-        self.nodes.push(Some(node));
-        index
-    }
-
-    fn get_node_ref(&self, index: GenericUpdatedNodeId) -> &UpdatedTrieStorageNodeWithSize {
-        self.nodes.get(index).expect(INVALID_STORAGE_HANDLE).as_ref().expect(INVALID_STORAGE_HANDLE)
-    }
-
-    fn store_value(&mut self, value: GenericTrieValue) -> ValueHandle {
-        let GenericTrieValue::MemtrieAndDisk(value) = value else {
-            unimplemented!(
-                "NodesStorage for Trie doesn't support value {value:?} \
-                because disk updates must be generated."
-            );
-        };
-
-        let value_len = value.len();
-        self.values.push(Some(value));
-        ValueHandle::InMemory(StorageValueHandle(self.values.len() - 1, value_len))
-    }
-
-    fn delete_value(&mut self, value: ValueHandle) -> Result<(), StorageError> {
-        match value {
-            ValueHandle::HashAndSize(value) => {
-                self.trie.internal_retrieve_trie_node(&value.hash, true, true)?;
-                self.refcount_changes.subtract(value.hash, 1);
-            }
-            ValueHandle::InMemory(_) => {
-                // Do nothing. Values which were just inserted were not
-                // refcounted yet.
-            }
-        }
         Ok(())
     }
 }
@@ -568,626 +278,6 @@ impl<'a, M: ArenaMemory> MemTrieUpdate<'a, M> {
         value: FlatStateValue,
     ) -> Result<(), StorageError> {
         self.generic_insert(0, key, GenericTrieValue::MemtrieOnly(value))
-    }
-}
-
-impl<'a, N, V, T> GenericTrieUpdateInsertDelete<'a, N, V> for T
-where
-    N: std::fmt::Debug,
-    V: std::fmt::Debug + HasValueLength,
-    T: GenericTrieUpdate<'a, N, V>,
-{
-}
-
-pub(crate) trait GenericTrieUpdateInsertDelete<'a, N, V>:
-    GenericTrieUpdateSquash<'a, N, V>
-where
-    N: std::fmt::Debug,
-    V: std::fmt::Debug + HasValueLength,
-{
-    fn calc_memory_usage_and_store(
-        &mut self,
-        node_id: GenericUpdatedNodeId,
-        children_memory_usage: u64,
-        new_node: GenericUpdatedTrieNode<N, V>,
-        old_child: Option<GenericUpdatedNodeId>,
-    ) {
-        let new_memory_usage =
-            children_memory_usage.saturating_add(new_node.memory_usage_direct()).saturating_sub(
-                old_child.map(|child| self.get_node_ref(child).memory_usage).unwrap_or_default(),
-            );
-        self.place_node_at(
-            node_id,
-            GenericUpdatedTrieNodeWithSize { node: new_node, memory_usage: new_memory_usage },
-        );
-    }
-
-    /// Insertion logic. We descend from the root down to whatever node corresponds to
-    /// the inserted value. We would need to split, modify, or transform nodes along
-    /// the way to achieve that. This takes care of refcounting changes for existing
-    /// nodes as well as values, but will not yet increment refcount for any newly
-    /// created nodes - that's done at the end.
-    fn generic_insert(
-        &mut self,
-        mut node_id: GenericUpdatedNodeId,
-        key: &[u8],
-        value: GenericTrieValue,
-    ) -> Result<(), StorageError> {
-        let mut partial = NibbleSlice::new(key);
-        // Path to the key being inserted.
-        // Needed to recompute memory usages in the end.
-        let mut path = Vec::new();
-
-        loop {
-            path.push(node_id);
-            // Take out the current node; we'd have to change it no matter what.
-            let GenericUpdatedTrieNodeWithSize { node, memory_usage } = self.take_node(node_id);
-            let children_memory_usage = memory_usage.saturating_sub(node.memory_usage_direct());
-
-            match node {
-                GenericUpdatedTrieNode::Empty => {
-                    // There was no node here, create a new leaf.
-                    let value_handle = self.store_value(value);
-                    let node = GenericUpdatedTrieNode::Leaf {
-                        extension: partial.encoded(true).into_vec().into_boxed_slice(),
-                        value: value_handle,
-                    };
-                    let memory_usage = node.memory_usage_direct();
-                    self.place_node_at(
-                        node_id,
-                        GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                    );
-                    break;
-                }
-                GenericUpdatedTrieNode::Branch { children, value: old_value } => {
-                    if partial.is_empty() {
-                        // This branch node is exactly where the value should be added.
-                        if let Some(value) = old_value {
-                            self.delete_value(value)?;
-                        }
-                        let value_handle = self.store_value(value);
-                        let node =
-                            GenericUpdatedTrieNode::Branch { children, value: Some(value_handle) };
-                        let memory_usage = children_memory_usage + node.memory_usage_direct();
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                        );
-                        break;
-                    } else {
-                        // Continue descending into the branch, possibly adding a new child.
-                        let mut new_children = children;
-                        let child = &mut new_children[partial.at(0) as usize];
-                        let new_node_id = match child.take() {
-                            Some(node_id) => self.ensure_updated(node_id)?,
-                            None => self.place_node(GenericUpdatedTrieNodeWithSize::empty()),
-                        };
-                        *child = Some(GenericNodeOrIndex::Updated(new_node_id));
-                        self.calc_memory_usage_and_store(
-                            node_id,
-                            children_memory_usage,
-                            GenericUpdatedTrieNode::Branch {
-                                children: new_children,
-                                value: old_value,
-                            },
-                            Some(new_node_id),
-                        );
-                        node_id = new_node_id;
-                        partial = partial.mid(1);
-                        continue;
-                    }
-                }
-                GenericUpdatedTrieNode::Leaf { extension, value: old_value } => {
-                    let existing_key = NibbleSlice::from_encoded(&extension).0;
-                    let common_prefix = partial.common_prefix(&existing_key);
-                    if common_prefix == existing_key.len() && common_prefix == partial.len() {
-                        // We're at the exact leaf. Rewrite the value at this leaf.
-                        self.delete_value(old_value)?;
-                        let value_handle = self.store_value(value);
-                        let node = GenericUpdatedTrieNode::Leaf { extension, value: value_handle };
-                        let memory_usage = node.memory_usage_direct();
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                        );
-                        break;
-                    } else if common_prefix == 0 {
-                        // Convert the leaf to an equivalent branch. We are not adding
-                        // the new branch yet; that will be done in the next iteration.
-                        let mut children = Box::<[_; 16]>::default();
-                        let children_memory_usage;
-                        let branch_node = if existing_key.is_empty() {
-                            // Existing key being empty means the old value now lives at the branch.
-                            children_memory_usage = 0;
-                            GenericUpdatedTrieNode::Branch { children, value: Some(old_value) }
-                        } else {
-                            let branch_idx = existing_key.at(0) as usize;
-                            let new_extension = existing_key.mid(1).encoded(true).into_vec();
-                            let new_node = GenericUpdatedTrieNode::Leaf {
-                                extension: new_extension.into_boxed_slice(),
-                                value: old_value,
-                            };
-                            let memory_usage = new_node.memory_usage_direct();
-                            children_memory_usage = memory_usage;
-                            let new_node_id = self.place_node(GenericUpdatedTrieNodeWithSize {
-                                node: new_node,
-                                memory_usage,
-                            });
-                            children[branch_idx] = Some(GenericNodeOrIndex::Updated(new_node_id));
-                            GenericUpdatedTrieNode::Branch { children, value: None }
-                        };
-                        let memory_usage =
-                            branch_node.memory_usage_direct() + children_memory_usage;
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node: branch_node, memory_usage },
-                        );
-                        path.pop();
-                        continue;
-                    } else {
-                        // Split this leaf into an extension plus a leaf, and descend into the leaf.
-                        let leaf_node = GenericUpdatedTrieNode::Leaf {
-                            extension: existing_key
-                                .mid(common_prefix)
-                                .encoded(true)
-                                .into_vec()
-                                .into_boxed_slice(),
-                            value: old_value,
-                        };
-                        let leaf_memory_usage = leaf_node.memory_usage_direct();
-                        let leaf_node_id = self.place_node(GenericUpdatedTrieNodeWithSize {
-                            node: leaf_node,
-                            memory_usage: leaf_memory_usage,
-                        });
-                        let extension_node = GenericUpdatedTrieNode::Extension {
-                            extension: partial
-                                .encoded_leftmost(common_prefix, false)
-                                .into_vec()
-                                .into_boxed_slice(),
-                            child: GenericNodeOrIndex::Updated(leaf_node_id),
-                        };
-                        let extension_memory_usage = extension_node.memory_usage_direct();
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize {
-                                node: extension_node,
-                                memory_usage: extension_memory_usage,
-                            },
-                        );
-                        node_id = leaf_node_id;
-                        partial = partial.mid(common_prefix);
-                        continue;
-                    }
-                }
-                GenericUpdatedTrieNode::Extension { extension, child: old_child, .. } => {
-                    let existing_key = NibbleSlice::from_encoded(&extension).0;
-                    let common_prefix = partial.common_prefix(&existing_key);
-                    if common_prefix == 0 {
-                        // Split Extension to Branch.
-                        let idx = existing_key.at(0);
-                        let child_memory_usage;
-                        let child = if existing_key.len() == 1 {
-                            child_memory_usage = children_memory_usage;
-                            old_child
-                        } else {
-                            let inner_child_node = GenericUpdatedTrieNode::Extension {
-                                extension: existing_key
-                                    .mid(1)
-                                    .encoded(false)
-                                    .into_vec()
-                                    .into_boxed_slice(),
-                                child: old_child,
-                            };
-                            child_memory_usage =
-                                children_memory_usage + inner_child_node.memory_usage_direct();
-                            let inner_child = GenericUpdatedTrieNodeWithSize {
-                                node: inner_child_node,
-                                memory_usage: child_memory_usage,
-                            };
-                            GenericNodeOrIndex::Updated(self.place_node(inner_child))
-                        };
-
-                        let mut children = Box::<[_; 16]>::default();
-                        children[idx as usize] = Some(child);
-                        let branch_node = GenericUpdatedTrieNode::Branch { children, value: None };
-                        let branch_memory_usage =
-                            branch_node.memory_usage_direct() + child_memory_usage;
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize {
-                                node: branch_node,
-                                memory_usage: branch_memory_usage,
-                            },
-                        );
-                        // Start over from the same position.
-                        path.pop();
-                        continue;
-                    } else if common_prefix == existing_key.len() {
-                        // Dereference child and descend into it.
-                        let child_id = self.ensure_updated(old_child)?;
-                        let node = GenericUpdatedTrieNode::Extension {
-                            extension,
-                            child: GenericNodeOrIndex::Updated(child_id),
-                        };
-                        let memory_usage = node.memory_usage_direct();
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                        );
-                        node_id = child_id;
-                        partial = partial.mid(common_prefix);
-                        continue;
-                    } else {
-                        // Partially shared prefix. Convert to shorter extension and descend into it.
-                        // On the next step, branch will be created.
-                        let inner_child_node = GenericUpdatedTrieNode::Extension {
-                            extension: existing_key
-                                .mid(common_prefix)
-                                .encoded(false)
-                                .into_vec()
-                                .into_boxed_slice(),
-                            child: old_child,
-                        };
-                        let inner_child_memory_usage =
-                            children_memory_usage + inner_child_node.memory_usage_direct();
-                        let inner_child_node_id = self.place_node(GenericUpdatedTrieNodeWithSize {
-                            node: inner_child_node,
-                            memory_usage: inner_child_memory_usage,
-                        });
-                        let child_node = GenericUpdatedTrieNode::Extension {
-                            extension: existing_key
-                                .encoded_leftmost(common_prefix, false)
-                                .into_vec()
-                                .into_boxed_slice(),
-                            child: GenericNodeOrIndex::Updated(inner_child_node_id),
-                        };
-                        let memory_usage = child_node.memory_usage_direct();
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node: child_node, memory_usage },
-                        );
-                        node_id = inner_child_node_id;
-                        partial = partial.mid(common_prefix);
-                        continue;
-                    }
-                }
-            }
-        }
-
-        for i in (0..path.len() - 1).rev() {
-            let node_id = path.get(i).unwrap();
-            let child_id = path.get(i + 1).unwrap();
-            let child_memory_usage = self.get_node_ref(*child_id).memory_usage;
-            let mut node = self.take_node(*node_id);
-            node.memory_usage += child_memory_usage;
-            self.place_node_at(*node_id, node);
-        }
-
-        Ok(())
-    }
-
-    /// Deletes a key from the trie.
-    ///
-    /// This will go down from the root of the trie to supposed location of the
-    /// key, deleting it if found. It will also keep the trie structure
-    /// consistent by changing the types of any nodes along the way.
-    ///
-    /// Deleting a non-existent key is allowed, and is a no-op.
-    fn generic_delete(
-        &mut self,
-        mut node_id: GenericUpdatedNodeId,
-        key: &[u8],
-    ) -> Result<(), StorageError> {
-        let mut partial = NibbleSlice::new(key);
-        // Path to find the key to delete.
-        // Needed to squash nodes and recompute memory usages in the end.
-        let mut path = vec![];
-        let mut key_deleted = true;
-
-        loop {
-            path.push(node_id);
-            let GenericUpdatedTrieNodeWithSize { node, memory_usage } = self.take_node(node_id);
-            let children_memory_usage = memory_usage.saturating_sub(node.memory_usage_direct());
-
-            match node {
-                GenericUpdatedTrieNode::Empty => {
-                    // Nothing to delete.
-                    self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize::empty());
-                    key_deleted = false;
-                    break;
-                }
-                GenericUpdatedTrieNode::Leaf { extension, value } => {
-                    if NibbleSlice::from_encoded(&extension).0 == partial {
-                        self.delete_value(value)?;
-                        self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize::empty());
-                        break;
-                    } else {
-                        // Key being deleted doesn't exist. Put the leaf back.
-                        let node = GenericUpdatedTrieNode::Leaf { extension, value };
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                        );
-                        key_deleted = false;
-                        break;
-                    }
-                }
-                GenericUpdatedTrieNode::Branch { mut children, value } => {
-                    if partial.is_empty() {
-                        if value.is_none() {
-                            // Key being deleted doesn't exist.
-                            let node = GenericUpdatedTrieNode::Branch { children, value };
-                            self.place_node_at(
-                                node_id,
-                                GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                            );
-                            key_deleted = false;
-                            break;
-                        };
-                        self.delete_value(value.unwrap())?;
-                        self.calc_memory_usage_and_store(
-                            node_id,
-                            children_memory_usage,
-                            GenericUpdatedTrieNode::Branch { children, value: None },
-                            None,
-                        );
-                        // if needed, branch will be squashed at the end of the function.
-                        break;
-                    } else {
-                        let child = &mut children[partial.at(0) as usize];
-                        let old_child_id = match child.take() {
-                            Some(node_id) => node_id,
-                            None => {
-                                // Key being deleted doesn't exist.
-                                let node = GenericUpdatedTrieNode::Branch { children, value };
-                                self.place_node_at(
-                                    node_id,
-                                    GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                                );
-                                key_deleted = false;
-                                break;
-                            }
-                        };
-                        let new_child_id = self.ensure_updated(old_child_id)?;
-                        *child = Some(GenericNodeOrIndex::Updated(new_child_id));
-                        self.calc_memory_usage_and_store(
-                            node_id,
-                            children_memory_usage,
-                            GenericUpdatedTrieNode::Branch { children, value },
-                            Some(new_child_id),
-                        );
-
-                        node_id = new_child_id;
-                        partial = partial.mid(1);
-                        continue;
-                    }
-                }
-                GenericUpdatedTrieNode::Extension { extension, child } => {
-                    let (common_prefix, existing_len) = {
-                        let extension_nibbles = NibbleSlice::from_encoded(&extension).0;
-                        (extension_nibbles.common_prefix(&partial), extension_nibbles.len())
-                    };
-                    if common_prefix == existing_len {
-                        let new_child_id = self.ensure_updated(child)?;
-                        self.calc_memory_usage_and_store(
-                            node_id,
-                            children_memory_usage,
-                            GenericUpdatedTrieNode::Extension {
-                                extension,
-                                child: GenericNodeOrIndex::Updated(new_child_id),
-                            },
-                            Some(new_child_id),
-                        );
-
-                        node_id = new_child_id;
-                        partial = partial.mid(existing_len);
-                        continue;
-                    } else {
-                        // Key being deleted doesn't exist.
-                        let node = GenericUpdatedTrieNode::Extension { extension, child };
-                        self.place_node_at(
-                            node_id,
-                            GenericUpdatedTrieNodeWithSize { node, memory_usage },
-                        );
-                        key_deleted = false;
-                        break;
-                    }
-                }
-            }
-        }
-
-        // Now we recompute memory usage and possibly squash nodes to keep the
-        // trie structure unique.
-        let mut child_memory_usage = 0;
-        for node_id in path.into_iter().rev() {
-            // First, recompute memory usage, emulating the recursive descent.
-            let GenericUpdatedTrieNodeWithSize { node, mut memory_usage } = self.take_node(node_id);
-            memory_usage += child_memory_usage;
-            self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize { node, memory_usage });
-
-            // Then, squash node to ensure unique trie structure, changing its
-            // type if needed. If `key_deleted` is false, trie structure is
-            // untouched.
-            if key_deleted {
-                self.squash_node(node_id)?;
-            }
-
-            child_memory_usage = self.get_node_ref(node_id).memory_usage;
-        }
-
-        Ok(())
-    }
-}
-
-impl<'a, N, V, T> GenericTrieUpdateSquash<'a, N, V> for T
-where
-    N: std::fmt::Debug,
-    V: std::fmt::Debug + HasValueLength,
-    T: GenericTrieUpdate<'a, N, V>,
-{
-}
-
-pub(crate) trait GenericTrieUpdateSquash<'a, N, V>: GenericTrieUpdate<'a, N, V>
-where
-    N: std::fmt::Debug,
-    V: std::fmt::Debug + HasValueLength,
-{
-    /// When we delete keys, it may be necessary to change types of some nodes,
-    /// in order to keep the trie structure unique. For example, if a branch
-    /// had two children, but after deletion ended up with one child and no
-    /// value, it must be converted to an extension node. Or, if an extension
-    /// node ended up having a child which is also an extension node, they must
-    /// be combined into a single extension node. This function takes care of
-    /// all these cases for a single node.
-    ///
-    /// To restructure trie correctly, this function must be called in
-    /// post-order traversal for every modified node. It may be proven by
-    /// induction on subtrees.
-    /// For single key removal, it is called for every node on the path from
-    /// the leaf to the root.
-    /// For range removal, it is called in the end of recursive range removal
-    /// function, which is the definition of post-order traversal.
-    fn squash_node(&mut self, node_id: GenericUpdatedNodeId) -> Result<(), StorageError> {
-        let GenericUpdatedTrieNodeWithSize { node, memory_usage } = self.take_node(node_id);
-        match node {
-            GenericUpdatedTrieNode::Empty => {
-                // Empty node will be absorbed by its parent node, so defer that.
-                self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize::empty());
-            }
-            GenericUpdatedTrieNode::Leaf { .. } => {
-                // It's impossible that we would squash a leaf node, because if we
-                // had deleted a leaf it would become Empty instead.
-                unreachable!();
-            }
-            GenericUpdatedTrieNode::Branch { mut children, value } => {
-                // Remove any children that are now empty (removed).
-                for child in children.iter_mut() {
-                    if let Some(GenericNodeOrIndex::Updated(child_node_id)) = child {
-                        if let GenericUpdatedTrieNode::Empty =
-                            self.get_node_ref(*child_node_id).node
-                        {
-                            *child = None;
-                        }
-                    }
-                }
-                let num_children = children.iter().filter(|node| node.is_some()).count();
-                if num_children == 0 {
-                    match value {
-                        None => {
-                            self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize::empty())
-                        }
-                        Some(value) => {
-                            // Branch with zero children and a value becomes leaf.
-                            let leaf_node = GenericUpdatedTrieNode::Leaf {
-                                extension: NibbleSlice::new(&[])
-                                    .encoded(true)
-                                    .into_vec()
-                                    .into_boxed_slice(),
-                                value,
-                            };
-                            let memory_usage = leaf_node.memory_usage_direct();
-                            self.place_node_at(
-                                node_id,
-                                GenericUpdatedTrieNodeWithSize { node: leaf_node, memory_usage },
-                            );
-                        }
-                    }
-                } else if num_children == 1 && value.is_none() {
-                    // Branch with 1 child but no value becomes extension.
-                    let (idx, child) = children
-                        .into_iter()
-                        .enumerate()
-                        .find_map(|(idx, node)| node.map(|node| (idx, node)))
-                        .unwrap();
-                    let extension = NibbleSlice::new(&[(idx << 4) as u8])
-                        .encoded_leftmost(1, false)
-                        .into_vec()
-                        .into_boxed_slice();
-                    self.extend_child(node_id, extension, child)?;
-                } else {
-                    // Branch with more than 1 children stays branch.
-                    self.place_node_at(
-                        node_id,
-                        GenericUpdatedTrieNodeWithSize {
-                            node: GenericUpdatedTrieNode::Branch { children, value },
-                            memory_usage,
-                        },
-                    );
-                }
-            }
-            GenericUpdatedTrieNode::Extension { extension, child } => {
-                self.extend_child(node_id, extension, child)?;
-            }
-        }
-        Ok(())
-    }
-
-    // Creates an extension node at `node_id`, but squashes the extension node according to
-    // its child; e.g. if the child is a leaf, the whole node becomes a leaf.
-    fn extend_child(
-        &mut self,
-        // The node being squashed.
-        node_id: GenericUpdatedNodeId,
-        // The current extension.
-        extension: Box<[u8]>,
-        // The current child.
-        child_id: GenericNodeOrIndex<N>,
-    ) -> Result<(), StorageError> {
-        let child_id = self.ensure_updated(child_id)?;
-        let GenericUpdatedTrieNodeWithSize { node, memory_usage } = self.take_node(child_id);
-        let child_child_memory_usage = memory_usage.saturating_sub(node.memory_usage_direct());
-        match node {
-            GenericUpdatedTrieNode::Empty => {
-                self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize::empty());
-            }
-            // If the child is a leaf (which could happen if a branch node lost
-            // all its branches and only had a value left, or is left with only
-            // one branch and that was squashed to a leaf).
-            GenericUpdatedTrieNode::Leaf { extension: child_extension, value } => {
-                let child_extension = NibbleSlice::from_encoded(&child_extension).0;
-                let extension = NibbleSlice::from_encoded(&extension)
-                    .0
-                    .merge_encoded(&child_extension, true)
-                    .into_vec()
-                    .into_boxed_slice();
-                let node = GenericUpdatedTrieNode::Leaf { extension, value };
-                let memory_usage = node.memory_usage_direct();
-                self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize { node, memory_usage });
-            }
-            // If the child is a branch, there's nothing to squash.
-            child_node @ GenericUpdatedTrieNode::Branch { .. } => {
-                self.place_node_at(
-                    child_id,
-                    GenericUpdatedTrieNodeWithSize { node: child_node, memory_usage },
-                );
-                let node = GenericUpdatedTrieNode::Extension {
-                    extension,
-                    child: GenericNodeOrIndex::Updated(child_id),
-                };
-                let memory_usage = memory_usage + node.memory_usage_direct();
-                self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize { node, memory_usage });
-            }
-            // If the child is an extension (which could happen if a branch node
-            // is left with only one branch), join the two extensions into one.
-            GenericUpdatedTrieNode::Extension {
-                extension: child_extension,
-                child: inner_child,
-            } => {
-                let child_extension = NibbleSlice::from_encoded(&child_extension).0;
-                let merged_extension = NibbleSlice::from_encoded(&extension)
-                    .0
-                    .merge_encoded(&child_extension, false)
-                    .into_vec()
-                    .into_boxed_slice();
-                let node = GenericUpdatedTrieNode::Extension {
-                    extension: merged_extension,
-                    child: inner_child,
-                };
-                let memory_usage = node.memory_usage_direct() + child_child_memory_usage;
-                self.place_node_at(node_id, GenericUpdatedTrieNodeWithSize { node, memory_usage });
-            }
-        }
-        Ok(())
     }
 }
